Create README.md
A committed professional with a decade of experience in management and operations, who is presently pursuing a Master of Science in Financial Technology and obtaining a certificate in Cybersecurity to improve their technical knowledge and professional proficiency.

Proficient in the application of data analytics and the utilization of sophisticated tools such as Tableau to facilitate strategic development and process optimization. Expertise in the integration of technological solutions and the utilization of HRIS systems to resolve organizational challenges.

The primary objective is to transition into a career in cybersecurity, leveraging a robust foundation in data-driven decision-making and developing proficiency in cybersecurity. The aim is to assist in developing innovative organizations by addressing intricate security challenges, protecting sensitive information, and instituting effective risk management strategies.