Update README.md
My name is Frida Desronvil. I am a dedicated and driven professional with a strong passion for technology and security. With a background in management and a current focus on pursuing a Master of Science in Financial Technology, I am expanding my expertise by working towards a certificate in Cybersecurity. I enjoy solving complex problems and leveraging data analytics to optimize processes and systems. My goal is to transition into the cybersecurity field, where I can apply my growing knowledge to address security challenges, safeguard sensitive information, and contribute to effective risk management strategies. The combination of my academic studies and cybersecurity certification will help me become a valuable asset in the field and prepare me for an impactful career in cybersecurity.